# docs/api/openapi_v001.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Managed Identity Wallets API",
    "description": "Managed Identity Wallets API",
    "termsOfService": "https://www.eclipse.org/legal/termsofuse.php",
    "contact": {
      "name": "Eclipse Tractus-X",
      "url": "https://projects.eclipse.org/projects/automotive.tractusx",
      "email": "tractusx-dev@eclipse.org"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://github.com/eclipse-tractusx/managed-identity-wallets/blob/develop/LICENSE"
    },
    "version": "0.0.1"
  },
  "security": [
    {"Authenticate using access_token": []},
    {"sts_token": []}
  ],
  "servers": [],
  "paths": {
...
    "/api/credentials/validation": {
      "post": {
        "tags": [
          "Verifiable Credential - Validation"
        ],
        "summary": "Validate Verifiable Credentials",
        "description": "Permission: **view_wallets** OR **view_wallet** \n\n Validate Verifiable Credentials",
        "operationId": "credentialsValidation",
        "parameters": [
          {
            "name": "withCredentialExpiryDate",
            "in": "query",
            "description": "Check expiry of VC",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialVerificationRequest"
              },
              "examples": {
                "Validate credential in JSON-LD format": {
                  "description": "Validate credential in JSON-LD format",
                  "value": {
                    "@context": [
                      "https://www.w3.org/2018/credentials/v1",
                      "https://catenax-ng.github.io/product-core-schemas/businessPartnerData.json",
                      "https://w3id.org/security/suites/jws-2020/v1"
                    ],
                    "id": "did:web:localhost:BPNL000000000000#f73e3631-ba87-4a03-bea3-b28700056879",
                    "type": [
                      "VerifiableCredential",
                      "BpnCredential"
                    ],
                    "issuer": "did:web:localhost:BPNL000000000000",
                    "issuanceDate": "2023-07-19T09:11:34Z",
                    "expirationDate": "2024-12-31T18:30:00Z",
                    "credentialSubject": [
                      {
                        "bpn": "BPNL000000000000",
                        "id": "did:web:localhost:BPNL000000000000",
                        "type": "BpnCredential"
                      }
                    ],
                    "proof": {
                      "created": "2023-07-19T09:11:39Z",
                      "jws": "eyJhbGciOiJFZERTQSJ9..fdn2qU85auOltdHDLdHI7sJVV1ZPdftpiXd_ndXN0dFgSDWiIrScdD03wtvKLq_H-shQWfh2RYeMmrlEzAhfDw",
                      "proofPurpose": "proofPurpose",
                      "type": "JsonWebSignature2020",
                      "verificationMethod": "did:web:localhost:BPNL000000000000#"
                    }
                  }
                },
                "Validate credential in JWT format": {
                  "description": "Validate credential in JWT format",
                  "value": {
                    "jwt": "eyJraWQiOiJkaWQ6d2ViOmFmODgtMjAzLTEyOS0yMTMtMTA3Lm5ncm9rLWZyZWUuYXBwOkJQTkwwMDAwMDAwMDAwMDAjOGYyZWU5ZDItYTM2Yy00MTM4LWJlMWYtYjZmZWZiNmY4MDI0IiwidHlwIjoiSldUIiwiYWxnIjoiRWREU0EifQ.eyJpc3MiOiJkaWQ6d2ViOmFmODgtMjAzLTEyOS0yMTMtMTA3Lm5ncm9rLWZyZWUuYXBwOkJQTkwwMDAwMDAwMDAwMDAiLCJzdWIiOiJkaWQ6d2ViOmFmODgtMjAzLTEyOS0yMTMtMTA3Lm5ncm9rLWZyZWUuYXBwOkJQTkwwMDAwMDAwMDAwMTEiLCJleHAiOjE3MzU2Njk4MDAsInZjIjp7IkBjb250ZXh0IjpbImh0dHBzOi8vd3d3LnczLm9yZy8yMDE4L2NyZWRlbnRpYWxzL3YxIiwiaHR0cHM6Ly9jb2Zpbml0eS14LmdpdGh1Yi5pby9zY2hlbWEtcmVnaXN0cnkvdjEuMS9Vc2VDYXNlVkMuanNvbiIsImh0dHBzOi8vdzNpZC5vcmcvc2VjdXJpdHkvc3VpdGVzL2p3cy0yMDIwL3YxIl0sImlkIjoiZGlkOndlYjphZjg4LTIwMy0xMjktMjEzLTEwNy5uZ3Jvay1mcmVlLmFwcDpCUE5MMDAwMDAwMDAwMDAwI2Q4Y2ZjZDBiLWY0NGQtNDVkMC05OGEzLTA4ZDZkNmU5Y2E5NSIsInR5cGUiOlsiVmVyaWZpYWJsZUNyZWRlbnRpYWwiLCJVc2VDYXNlRnJhbWV3b3JrQ29uZGl0aW9uIl0sImlzc3VlciI6ImRpZDp3ZWI6YWY4OC0yMDMtMTI5LTIxMy0xMDcubmdyb2stZnJlZS5hcHA6QlBOTDAwMDAwMDAwMDAwMCIsImNyZWRlbnRpYWxTdWJqZWN0IjpbeyJob2xkZXJJZGVudGlmaWVyIjoiQlBOTDAwMDAwMDAwMDAxMSIsImlkIjoiZGlkOndlYjphZjg4LTIwMy0xMjktMjEzLTEwNy5uZ3Jvay1mcmVlLmFwcDpCUE5MMDAwMDAwMDAwMDExIiwidHlwZSI6IkJlaGF2aW9yVHdpbkNyZWRlbnRpYWwiLCJjb250cmFjdFRlbXBsYXRlIjoiaHR0cHM6Ly9wdWJsaWMuY2F0ZW5hLXgub3JnL2NvbnRyYWN0cy90cmFjZWFiaWx0eS52MS5wZGYiLCJjb250cmFjdFZlcnNpb24iOiIxLjAuMCJ9XSwiY3JlZGVudGlhbFN0YXR1cyI6bnVsbCwiaXNzdWFuY2VEYXRlIjoiMjAyNC0wMi0wOFQxNDowMjo1M1oiLCJleHBpcmF0aW9uRGF0ZSI6IjIwMjQtMTItMzFUMTg6MzA6MDBaIn0sImp0aSI6IjliYWFhMjIzLTAxMjctNDEyZS05NjZhLTA3ZTJmZGU4NGNlNCJ9.X3rkj8Gv4OD5nEaeFG5pSA-dogbcYA91YEPmHiKT4FhAiIr7QAdSEULGXHYOn8-eK0jSDHNdAxNYIK1UwYRsCA"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Validate Verifiable Credentials",
            "content": {
              "application/json": {
                "examples": {
                  "Verifiable Credentials without check expiry": {
                    "description": "Verifiable Credentials without check expiry",
                    "value": {
                      "valid": true,
                      "vc": {
                        "issuanceDate": "2023-07-19T09:11:34Z",
                        "credentialSubject": [
                          {
                            "bpn": "BPNL000000000000",
                            "id": "did:web:localhost:BPNL000000000000",
                            "type": "BpnCredential"
                          }
                        ],
                        "id": "did:web:localhost:BPNL000000000000#f73e3631-ba87-4a03-bea3-b28700056879",
                        "proof": {
                          "created": "2023-07-19T09:11:39Z",
                          "jws": "eyJhbGciOiJFZERTQSJ9..fdn2qU85auOltdHDLdHI7sJVV1ZPdftpiXd_ndXN0dFgSDWiIrScdD03wtvKLq_H-shQWfh2RYeMmrlEzAhfDw",
                          "proofPurpose": "proofPurpose",
                          "type": "JsonWebSignature2020",
                          "verificationMethod": "did:web:localhost:BPNL000000000000#"
                        },
                        "type": [
                          "VerifiableCredential",
                          "BpnCredential"
                        ],
                        "@context": [
                          "https://www.w3.org/2018/credentials/v1",
                          "https://catenax-ng.github.io/product-core-schemas/businessPartnerData.json",
                          "https://w3id.org/security/suites/jws-2020/v1"
                        ],
                        "issuer": "did:web:localhost:BPNL000000000000",
                        "expirationDate": "2024-12-31T18:30:00Z"
                      }
                    }
                  },
                  "Verifiable Credentials with check expiry": {
                    "description": "Verifiable Credentials with check expiry",
                    "value": {
                      "valid": true,
                      "validateExpiryDate": true,
                      "vc": {
                        "issuanceDate": "2023-07-19T09:11:34Z",
                        "credentialSubject": [
                          {
                            "bpn": "BPNL000000000000",
                            "id": "did:web:localhost:BPNL000000000000",
                            "type": "BpnCredential"
                          }
                        ],
                        "id": "did:web:localhost:BPNL000000000000#f73e3631-ba87-4a03-bea3-b28700056879",
                        "proof": {
                          "created": "2023-07-19T09:11:39Z",
                          "jws": "eyJhbGciOiJFZERTQSJ9..fdn2qU85auOltdHDLdHI7sJVV1ZPdftpiXd_ndXN0dFgSDWiIrScdD03wtvKLq_H-shQWfh2RYeMmrlEzAhfDw",
                          "proofPurpose": "proofPurpose",
                          "type": "JsonWebSignature2020",
                          "verificationMethod": "did:web:localhost:BPNL000000000000#"
                        },
                        "type": [
                          "VerifiableCredential",
                          "BpnCredential"
                        ],
                        "@context": [
                          "https://www.w3.org/2018/credentials/v1",
                          "https://catenax-ng.github.io/product-core-schemas/businessPartnerData.json",
                          "https://w3id.org/security/suites/jws-2020/v1"
                        ],
                        "issuer": "did:web:localhost:BPNL000000000000",
                        "expirationDate": "2024-12-31T18:30:00Z"
                      }
                    }
                  },
                  "Verifiable expired credentials with check expiry ": {
                    "description": "Verifiable expired credentials with check expiry ",
                    "value": {
                      "valid": false,
                      "validateExpiryDate": false,
                      "vc": {
                        "issuanceDate": "2023-07-19T09:11:34Z",
                        "credentialSubject": [
                          {
                            "bpn": "BPNL000000000000",
                            "id": "did:web:localhost:BPNL000000000000",
                            "type": "BpnCredential"
                          }
                        ],
                        "id": "did:web:localhost:BPNL000000000000#f73e3631-ba87-4a03-bea3-b28700056879",
                        "proof": {
                          "created": "2023-07-19T09:11:39Z",
                          "jws": "eyJhbGciOiJFZERTQSJ9..fdn2qU85auOltdHDLdHI7sJVV1ZPdftpiXd_ndXN0dFgSDWiIrScdD03wtvKLq_H-shQWfh2RYeMmrlEzAhfDw",
                          "proofPurpose": "proofPurpose",
                          "type": "JsonWebSignature2020",
                          "verificationMethod": "did:web:localhost:BPNL000000000000#"
                        },
                        "type": [
                          "VerifiableCredential",
                          "BpnCredential"
                        ],
                        "@context": [
                          "https://www.w3.org/2018/credentials/v1",
                          "https://catenax-ng.github.io/product-core-schemas/businessPartnerData.json",
                          "https://w3id.org/security/suites/jws-2020/v1"
                        ],
                        "issuer": "did:web:localhost:BPNL000000000000",
                        "expirationDate": "2022-12-31T18:30:00Z"
                      }
                    }
                  },
                  "Revocable Verifiable credentials with check expiry ": {
                    "description": "Revocable Verifiable credentials with check expiry ",
                    "value": {
                      "credentialStatus": "active",
                      "valid": true,
                      "validateExpiryDate": true,
                      "vc": {
                        "credentialSubject": [
                          {
                            "holderIdentifier": "BPNL000000000001",
                            "allowedVehicleBrands": [
                              "Audi",
                              "Abarth",
                              "Alfa Romeo",
                              "Chrysler"
                            ],
                            "id": "did:web:6e3e-203-129-213-107.ngrok-free.app:BPNL000000000001",
                            "activityType": "vehicleDismantle",
                            "type": "DismantlerCredential"
                          }
                        ],
                        "issuanceDate": "2024-01-05T05:42:53Z",
                        "id": "did:web:6e3e-203-129-213-107.ngrok-free.app:BPNL000000000000#8507aa50-b2a4-4532-8e45-f50e7654b23b",
                        "proof": {
                          "proofPurpose": "assertionMethod",
                          "verificationMethod": "did:web:6e3e-203-129-213-107.ngrok-free.app:BPNL000000000000#a39d8ccf-2a66-488d-bfec-916768082e91",
                          "type": "JsonWebSignature2020",
                          "created": "2024-01-05T05:42:53Z",
                          "jws": "eyJhbGciOiJFZERTQSJ9..15NdxA8L_Iw7Igxevm7YGMAQA-Kt6PMOpix6p0jaYHCtfQnTy3q61SDvsnsltGT6fzM90JOubOuig2WFy-GPDg"
                        },
                        "type": [
                          "VerifiableCredential",
                          "DismantlerCredential"
                        ],
                        "@context": [
                          "https://www.w3.org/2018/credentials/v1",
                          "https://cofinity-x.github.io/schema-registry/v1.1/DismantlerVC.json",
                          "https://w3id.org/security/suites/jws-2020/v1",
                          "https://w3id.org/vc/status-list/2021/v1"
                        ],
                        "issuer": "did:web:6e3e-203-129-213-107.ngrok-free.app:BPNL000000000000",
                        "credentialStatus": {
                          "id": "did:web:6e3e-203-129-213-107.ngrok-free.app:BPNL000000000000#0",
                          "statusPurpose": "revocation",
                          "statusListIndex": "0",
                          "statusListCredential": "https://ae96-203-129-213-107.ngrok-free.app/api/v1/revocations/credentials?issuerId=did:web:6e3e-203-129-213-107.ngrok-free.app:BPNL000000000000",
                          "type": "StatusList2021Entry"
                        },
                        "expirationDate": "2024-12-31T18:30:00Z"
                      }
                    }
                  },
                  "Verifiable Credentials with invalid signature": {
                    "description": "Verifiable Credentials with invalid signature",
                    "value": {
                      "valid": false,
                      "vc": {
                        "@context": [
                          "https://www.w3.org/2018/credentials/v1",
                          "https://catenax-ng.github.io/product-core-schemas/businessPartnerData.json",
                          "https://w3id.org/security/suites/jws-2020/v1"
                        ],
                        "id": "did:web:localhost:BPNL000000000000#f73e3631-ba87-4a03-bea3-b28700056879",
                        "type": [
                          "VerifiableCredential",
                          "BpnCredential"
                        ],
                        "issuer": "did:web:localhost:BPNL000000000000",
                        "expirationDate": "2024-12-31T18:30:00Z",
                        "issuanceDate": "2023-07-19T09:11:34Z",
                        "credentialSubject": [
                          {
                            "bpn": "BPNL000000000000",
                            "id": "did:web:localhost:BPNL000000000000",
                            "type": "BpnCredential"
                          }
                        ],
                        "proof": {
                          "created": "2023-07-19T09:11:39Z",
                          "jws": "eyJhbGciOiJFZERTQSJ9..fdn2qU85auOltdHDLdHI7sJVV1ZPdftpiXd_ndXN0dFgSDWiIrScdD03wtvKLq_H-shQWfh2RYeMmrlEzAhf",
                          "proofPurpose": "proofPurpose",
                          "type": "JsonWebSignature2020",
                          "verificationMethod": "did:web:localhost:BPNL000000000000#"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "The request could not be completed due to a failed authorization."
          },
          "403": {
            "description": "The request could not be completed due to a forbidden access"
          },
          "500": {
            "description": "Any other internal server error",
            "content": {
              "application/json": {
                "examples": {
                  "Internal server error": {
                    "description": "Internal server error",
                    "value": {
                      "type": "about:blank",
                      "title": "Error Title",
                      "status": 500,
                      "detail": "Error Details",
                      "instance": "API endpoint",
                      "properties": {
                        "timestamp": 1689762476720
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authenticate using access_token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CredentialVerificationRequest": {
        "type": "object",
        "properties": {
          "jwt": {
            "type": "string",
            "writeOnly": true
          },
          "vc": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            },
            "writeOnly": true
          },
          "empty": {
            "type": "boolean"
          }
        },
        "additionalProperties": {
          "type": "object"
        }
      }
    },
    "securitySchemes": {
      "Authenticate using access_token": {
        "type": "apiKey",
        "description": "**Bearer (apiKey)**\nJWT Authorization header using the Bearer scheme.\nEnter **Bearer** [space] and then your token in the text input below:\nExample: Bearer 12345abcdef\n",
        "name": "Authorization",
        "in": "header"
      },
      "sts_token": {
        "type": "apiKey",
        "description": "**STS token**\nJWT Authorization header.\nEnter your token in the text input below:\nExample: 12345abcdef\n",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}